smp: factor finding baseline sha into a job
If we factor out the process of finding a baseline SHA for the
Regression Detector into a separate job, we can impose a shorter
timeout for that step so that we don't have to wait 2 hours for it to
fail. This commit is an attempt to do that refactoring, but it hasn't
been checked as to whether or not it works, so the changes may require
debugging.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>